master: working on warnings on some systems. CMakeLists.txt: supppress all warnings in optional packages. cppad_for_tmb.cpp: fix warning when using boost.

+++ test_more/cppad_for_tmb/cppad_for_tmb.cpp
@@ -197,7 +197,7 @@ bool multi_thread_checkpoint(void)
 		v         = f.Forward(0, x);
 		//
 		// this assigment has false sharing; i.e., will case cache resets
-		y[thread] = v[0];
+		y[ size_t(thread) ] = v[0];
 	}
 
 	// check the results
